🔨 Refactor: 공통부분 컴포넌트 분리
